ASE bake-off: NLA residual steering vs CodeSteer on CodeLlama-7B (report + H-R2/H-R4/H-R6)
Adds reports/2026-09-15_ase-bakeoff/REPORT.md — one line per steering
method (16 arms, case-weighted Pass@1, paired cluster-bootstrap CIs,
parse-rate confound) and the frozen-rule verdicts:
  H-R2a MATCH-CODESTEER, H-R2b SLICE-IRRELEVANT, H-R2c unreadable,
  H-R4 refuted, H-R6a MAP-NOT-BETTER, H-R6b CATEGORY-MEANING-INERT,
  H-R6c PROMPT-SUFFICES (conservative reading).

Also commits the ASE harness (nla/src/ase_*.py, configs, scripts,
tests), the kept bake-off numbers (nla/results/2026-09-15_ase_bakeoff/),
the 2026-09-09…09-15 log/nla-harness entries with thread + master
index updates, and the CHECKLIST.md move from docs/ to the project base.

+++ nla/src/local_av.py
@@ -56,7 +56,7 @@ class LocalAV:
         self.tokenizer = self._c.tokenizer
         self.cfg = self._c.cfg
         self.model = AutoModelForCausalLM.from_pretrained(
-            str(checkpoint_dir), torch_dtype=dtype, device_map=device).eval()
+            str(checkpoint_dir), dtype=dtype, device_map=device).eval()
         self.device = self.model.device
         self.n_no_tags = 0
